Initial14+cleaning
