+++ lib/home/widgets/custom_appbar.dart
@@ -37,51 +37,5 @@ customAppBar(List<Map<String, dynamic>> unidades) {
         ),
       ),
     ],
-    //   Padding(
-    //     padding: const EdgeInsets.only(right: 12.0),
-    //     child: IconButton(
-    //       icon: const Icon(Icons.compare_arrows_rounded),
-    //       tooltip: 'Data network list page',
-    //       onPressed: () async {
-    //         await Future.delayed(const Duration(milliseconds: 500));
-    //         Get.to(
-    //           () => RedeDados(
-    //             unidades: unidades,
-    //           ),
-    //         );
-    //       },
-    //     ),
-    //   ),
-    //   Padding(
-    //     padding: const EdgeInsets.only(right: 12.0),
-    //     child: IconButton(
-    //       icon: const Icon(Icons.router_outlined),
-    //       tooltip: 'Switches list page',
-    //       onPressed: () async {
-    //         await Future.delayed(const Duration(milliseconds: 500));
-    //         Get.to(
-    //           () => SwitchesPage(
-    //             unidades: unidades,
-    //           ),
-    //         );
-    //       },
-    //     ),
-    //   ),
-    //   Padding(
-    //     padding: const EdgeInsets.only(right: 12.0),
-    //     child: IconButton(
-    //       icon: const Icon(Icons.switch_access_shortcut),
-    //       tooltip: 'Onus list page',
-    //       onPressed: () async {
-    //         await Future.delayed(const Duration(milliseconds: 500));
-    //         Get.to(
-    //           () => OnusPage(
-    //             unidades: unidades,
-    //           ),
-    //         );
-    //       },
-    //     ),
-    //   ),
-    // ],
   );
 }
