+++ lib/onus/onus_page.dart
@@ -1,8 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
-import 'package:proprios/detail/detail_page.dart';
 import 'package:proprios/home/widgets/custom_appbar.dart';
 import 'package:proprios/home/widgets/custom_drawer.dart';
+import 'package:proprios/onus/controller/onus_controller.dart';
+import 'package:proprios/onus/widgets/custom_inkwell.dart';
 import 'package:proprios/search/search_page.dart';
 
 class OnusPageBinding extends Bindings {
@@ -15,71 +16,14 @@ class OnusPageBinding extends Bindings {
 class OnusPageController extends GetxController {
   @override
   void onInit() {
+    debugPrint('on Init');
     super.onInit();
   }
 
   @override
   void onReady() {
+    debugPrint('on Ready');
     super.onReady();
-  }
-}
-
-class OnusController extends GetxController {
-  late List<Map<String, dynamic>> mutableList;
-  final ipv4Regex = RegExp(
-      r'^((25[0-5]|2[0-4][0-9]|1[0-9][0-9]|[1-9]?[0-9])\.){3}(25[0-5]|2[0-4][0-9]|1[0-9][0-9]|[1-9]?[0-9])$');
-
-  //Sort ipv4 addresses
-  List<String> sortIPv4Addresses(List<String> ipAddresses) {
-    List<int> numericalAddresses = [];
-
-    // Convert IP addresses to numerical representations
-    for (String ip in ipAddresses) {
-      List<int> octets = ip.split('.').map(int.parse).toList();
-      int numericalAddress =
-          octets[0] << 24 | octets[1] << 16 | octets[2] << 8 | octets[3];
-      numericalAddresses.add(numericalAddress);
-    }
-
-    // Sort the numerical representations
-    numericalAddresses.sort();
-
-    // Convert back to IP addresses
-    return numericalAddresses.map((numericalAddress) {
-      int a = (numericalAddress >> 24) & 0xFF;
-      int b = (numericalAddress >> 16) & 0xFF;
-      int c = (numericalAddress >> 8) & 0xFF;
-      int d = numericalAddress & 0xFF;
-      return "$a.$b.$c.$d";
-    }).toList();
-  }
-
-  bool isValidIPv4(String ipAddress) {
-    return ipv4Regex.hasMatch(ipAddress);
-  }
-
-  List<String> orderedList = [];
-
-  void updateList(List<Map<String, dynamic>> newList) {
-    //create a mutable copy of proprios
-    mutableList = List<Map<String, dynamic>>.from(newList);
-
-    //initial list of ips
-    List<String> rd = [];
-
-    //traverse List<Map<String, dynamic>> proprios
-    for (var proprio in mutableList) {
-      String ip = proprio['ipgerenciaonu'];
-
-      List<String> newip = ip.split('/');
-      for (var x in newip) {
-        if (isValidIPv4(x)) {
-          rd.add(x);
-        }
-      }
-    }
-
-    orderedList = sortIPv4Addresses(rd);
   }
 }
 
@@ -92,7 +36,10 @@ class OnusPage extends GetView<OnusPageController> {
 
   @override
   Widget build(BuildContext context) {
+    //initialize ahead of time (use of search in appBar)
+    // ignore: unused_local_variable
     final MyController mycontroller = Get.put(MyController());
+
     final scontroller = Get.put(OnusController());
     scontroller.updateList(unidades);
     return Scaffold(
@@ -106,30 +53,8 @@ class OnusPage extends GetView<OnusPageController> {
             mainAxisSpacing: 2),
         itemCount: scontroller.orderedList.length,
         itemBuilder: (BuildContext context, int index) {
-          return InkWell(
-            onTap: () {
-              final matchingIndex = unidades.indexWhere(
-                (item) => item['ipgerenciaonu']
-                    .contains(scontroller.orderedList[index]),
-              );
-              Get.to(
-                DetailPage(
-                  index: matchingIndex,
-                  unidades: unidades,
-                ),
-              );
-            },
-            child: Container(
-              color: Colors.greenAccent,
-              child: Center(
-                // tileColor: Colors.greenAccent,
-                child: Text(
-                  scontroller.orderedList[index],
-                  style: TextStyle(fontSize: 11.0, fontWeight: FontWeight.bold),
-                ),
-              ),
-            ),
-          );
+          return CustomInkwell(
+              unidades: unidades, index: index, scontroller: scontroller);
         },
       ),
     );
