+++ lib/utils/code_pieces.dart/codes_unused.dart
@@ -1,0 +1,67 @@
+//FloatingButton on home Scaffold
+   // floatingActionButton: FloatingActionButton(
+      //   backgroundColor: Colors.lightGreenAccent,
+      //   tooltip: 'Search...',
+      //   onPressed: () async {
+      //     await Future.delayed(const Duration(milliseconds: 500));
+      //     Get.to(
+      //       () => SearchPage(
+      //         unidades: unidades,
+      //       ),
+      //     );
+      //     // Get.to(SearchPage());
+      //   },
+      //   child: Icon(
+      //     Icons.search,
+      //     color: Colors.green,
+      //   ),
+      // ),
+      // floatingActionButtonLocation: FloatingActionButtonLocation.centerFloat,
+
+//not used appbar actions on home page 
+ //   Padding(
+    //     padding: const EdgeInsets.only(right: 12.0),
+    //     child: IconButton(
+    //       icon: const Icon(Icons.compare_arrows_rounded),
+    //       tooltip: 'Data network list page',
+    //       onPressed: () async {
+    //         await Future.delayed(const Duration(milliseconds: 500));
+    //         Get.to(
+    //           () => RedeDados(
+    //             unidades: unidades,
+    //           ),
+    //         );
+    //       },
+    //     ),
+    //   ),
+    //   Padding(
+    //     padding: const EdgeInsets.only(right: 12.0),
+    //     child: IconButton(
+    //       icon: const Icon(Icons.router_outlined),
+    //       tooltip: 'Switches list page',
+    //       onPressed: () async {
+    //         await Future.delayed(const Duration(milliseconds: 500));
+    //         Get.to(
+    //           () => SwitchesPage(
+    //             unidades: unidades,
+    //           ),
+    //         );
+    //       },
+    //     ),
+    //   ),
+    //   Padding(
+    //     padding: const EdgeInsets.only(right: 12.0),
+    //     child: IconButton(
+    //       icon: const Icon(Icons.switch_access_shortcut),
+    //       tooltip: 'Onus list page',
+    //       onPressed: () async {
+    //         await Future.delayed(const Duration(milliseconds: 500));
+    //         Get.to(
+    //           () => OnusPage(
+    //             unidades: unidades,
+    //           ),
+    //         );
+    //       },
+    //     ),
+    //   ),
+    // ],
